perf: 配置文件页放到单独组件
